src/llm_config.py
import os
import google.generativeai as genai
from google.generativeai.types import HarmCategory, HarmBlockThreshold

# Configure Gemini API
genai.configure(api_key=os.getenv("GOOGLE_API_KEY"))

# Safety settings for Gemini API
safety_settings = {
    HarmCategory.HARM_CATEGORY_HARASSMENT: HarmBlockThreshold.BLOCK_NONE,
    HarmCategory.HARM_CATEGORY_HATE_SPEECH: HarmBlockThreshold.BLOCK_NONE,
    HarmCategory.HARM_CATEGORY_SEXUALLY_EXPLICIT: HarmBlockThreshold.BLOCK_NONE,
    HarmCategory.HARM_CATEGORY_DANGEROUS_CONTENT: HarmBlockThreshold.BLOCK_NONE,
}

def get_model(model_name: str = "gemini-1.5-flash-latest") -> genai.GenerativeModel:
    """Get a configured Gemini model instance."""
    return genai.GenerativeModel(model_name) 