requirements.txt
pydantic
modal
python-dotenv
python-telegram-bot
telegramify-markdown
google-generativeai
anthropic
openai